requirements.txt
python-whois==0.7.3
beautifulsoup4==4.10.0
requests==2.26.0
lxml==4.6.4
pandas==1.3.4
tqdm==4.62.3
scikit-learn==1.0.1
xgboost==0.80